+++ docker/test_env/env_component_test.go
@@ -39,7 +39,7 @@ func TestEnvComponentPauseChaos(t *testing.T) {
 	require.NoError(t, err)
 
 	defaultChainCfg := GetDefaultChainConfig()
-	g := NewGethPoa([]string{network.Name}, &defaultChainCfg).
+	g := NewGethEth1([]string{network.Name}, &defaultChainCfg).
 		WithTestInstance(t)
 	_, err = g.StartContainer()
 	require.NoError(t, err)
