+++ docker/test_env/erigon_eth1.go
@@ -1,0 +1,181 @@
+package test_env
+
+import (
+	"bytes"
+	"fmt"
+	"html/template"
+	"os"
+	"strings"
+	"time"
+
+	"github.com/docker/docker/api/types/container"
+	"github.com/docker/docker/api/types/mount"
+	"github.com/google/uuid"
+	tc "github.com/testcontainers/testcontainers-go"
+	tcwait "github.com/testcontainers/testcontainers-go/wait"
+
+	"github.com/smartcontractkit/chainlink-testing-framework/logging"
+	"github.com/smartcontractkit/chainlink-testing-framework/mirror"
+	"github.com/smartcontractkit/chainlink-testing-framework/utils/templates"
+)
+
+func NewErigonEth1(networks []string, chainConfg *EthereumChainConfig, opts ...EnvComponentOption) (*Erigon, error) {
+	parts := strings.Split(defaultErigonEth1Image, ":")
+	g := &Erigon{
+		EnvComponent: EnvComponent{
+			ContainerName:    fmt.Sprintf("%s-%s", "erigon-eth1", uuid.NewString()[0:8]),
+			Networks:         networks,
+			ContainerImage:   parts[0],
+			ContainerVersion: parts[1],
+		},
+		chainConfg: chainConfg,
+		l:          logging.GetTestLogger(nil),
+	}
+	g.SetDefaultHooks()
+	for _, opt := range opts {
+		opt(&g.EnvComponent)
+	}
+	// if the internal docker repo is set then add it to the version
+	g.EnvComponent.ContainerImage = mirror.AddMirrorToImageIfSet(g.EnvComponent.ContainerImage)
+	return g, nil
+}
+
+func (g *Erigon) getEth1ContainerRequest() (*tc.ContainerRequest, error) {
+	keystoreDir, err := os.MkdirTemp("", "keystore")
+	if err != nil {
+		return nil, err
+	}
+
+	generatedData, err := generateKeystoreAndExtraData(keystoreDir)
+	if err != nil {
+		return nil, err
+	}
+
+	genesisJsonStr, err := templates.GenesisJsonTemplate{
+		ChainId:     fmt.Sprintf("%d", g.chainConfg.ChainID),
+		AccountAddr: []string{RootFundingAddr, generatedData.minerAccount.Address.Hex()},
+		Consensus:   templates.GethGenesisConsensus_Ethash,
+	}.String()
+	if err != nil {
+		return nil, err
+	}
+
+	initFile, err := os.CreateTemp("", "init.sh")
+	if err != nil {
+		return nil, err
+	}
+
+	initScriptContent, err := g.buildPowInitScript(generatedData.minerAccount.Address.Hex())
+	if err != nil {
+		return nil, err
+	}
+
+	_, err = initFile.WriteString(initScriptContent)
+	if err != nil {
+		return nil, err
+	}
+
+	genesisFile, err := os.CreateTemp("", "genesis_json")
+	if err != nil {
+		return nil, err
+	}
+	_, err = genesisFile.WriteString(genesisJsonStr)
+	if err != nil {
+		return nil, err
+	}
+	key1File, err := os.CreateTemp(keystoreDir, "key1")
+	if err != nil {
+		return nil, err
+	}
+	_, err = key1File.WriteString(RootFundingWallet)
+	if err != nil {
+		return nil, err
+	}
+
+	return &tc.ContainerRequest{
+		Name:          g.ContainerName,
+		Image:         g.GetImageWithVersion(),
+		Networks:      g.Networks,
+		ImagePlatform: "linux/x86_64",
+		ExposedPorts:  []string{NatPortFormat(DEFAULT_EVM_NODE_HTTP_PORT)},
+		WaitingFor: tcwait.ForAll(
+			tcwait.ForLog("Started P2P networking").
+				WithStartupTimeout(120*time.Second).
+				WithPollInterval(1*time.Second),
+			NewWebSocketStrategy(NatPort(DEFAULT_EVM_NODE_HTTP_PORT), g.l),
+		),
+		User: "0:0",
+		Entrypoint: []string{
+			"sh",
+			"/home/erigon/init.sh",
+		},
+		Files: []tc.ContainerFile{
+			{
+				HostFilePath:      initFile.Name(),
+				ContainerFilePath: "/root/init.sh",
+				FileMode:          0644,
+			},
+			{
+				HostFilePath:      genesisFile.Name(),
+				ContainerFilePath: "/root/genesis.json",
+				FileMode:          0644,
+			},
+			{
+				HostFilePath:      initFile.Name(),
+				ContainerFilePath: "/home/erigon/init.sh",
+				FileMode:          0744,
+			},
+		},
+		HostConfigModifier: func(hostConfig *container.HostConfig) {
+			hostConfig.Mounts = append(hostConfig.Mounts, mount.Mount{
+				Type:     mount.TypeBind,
+				Source:   keystoreDir,
+				Target:   "/root/.local/share/erigon/keystore/",
+				ReadOnly: false,
+			},
+			)
+		},
+		LifecycleHooks: []tc.ContainerLifecycleHooks{
+			{
+				PostStarts: g.PostStartsHooks,
+				PostStops:  g.PostStopsHooks,
+			},
+		},
+	}, nil
+}
+
+func (g *Erigon) buildPowInitScript(minerAddr string) (string, error) {
+	initTemplate := `#!/bin/bash
+	echo "Running erigon init"
+	erigon init /root/genesis.json
+	exit_code=$?
+	if [ $exit_code -ne 0 ]; then
+		echo "Erigon init failed with exit code $exit_code"
+		exit 1
+	fi
+
+	echo "Starting Erigon..."
+	erigon --http --http.api=eth,erigon,engine,web3,net,debug,trace,txpool,admin --http.addr=0.0.0.0 --http.corsdomain=* --http.vhosts=* --http.port={{.HttpPort}} --ws \
+	--allow-insecure-unlock  --nodiscover --networkid={{.ChainID}} --mine --miner.etherbase={{.MinerAddr}} --fakepow`
+
+	data := struct {
+		HttpPort  string
+		ChainID   int
+		MinerAddr string
+	}{
+		HttpPort:  DEFAULT_EVM_NODE_HTTP_PORT,
+		ChainID:   g.chainConfg.ChainID,
+		MinerAddr: minerAddr,
+	}
+
+	t, err := template.New("init").Parse(initTemplate)
+	if err != nil {
+		fmt.Println("Error parsing template:", err)
+		os.Exit(1)
+	}
+
+	var buf bytes.Buffer
+	err = t.Execute(&buf, data)
+
+	return buf.String(), err
+}
